requirements.txt
langchain>=0.2.12
langchain-openai>=0.1.22
langchain-community>=0.2.11
pydantic>=2.7
python-dotenv>=1.0
pandas>=2.2
openpyxl>=3.1
faiss-cpu>=1.8
pymupdf>=1.24
unstructured>=0.14
tiktoken>=0.7
scikit-learn>=1.4
rich>=13.7
fastapi>=0.111
uvicorn[standard]>=0.30
python-multipart>=0.0.9
gradio>=4.40
requests>=2.32
langchain-google-genai>=1.0.7
Pillow>=10.4
